+++ src/main/java/com/example/kp3coutsourcingproject/feed/repository/FeedRepository.java
@@ -1,6 +1,6 @@
-package com.example.kp3coutsourcingproject.timeline.repository;
+package com.example.kp3coutsourcingproject.feed.repository;
 
-import com.example.kp3coutsourcingproject.timeline.entity.Feed;
+import com.example.kp3coutsourcingproject.feed.entity.Feed;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
